% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ArdecheStMartin}
\alias{ArdecheStMartin}
\title{Simulated river discharge of the Ardeche river at Saint Martin d'Ardeche}
\format{
Two column dataframe:
#' \itemize{
\item time (POSIXct timestamp with 6-hourly resolution)
\item Q (6-houry mean discharge in cubic meter per second)
}
}
\source{
\url{https://data.jrc.ec.europa.eu/dataset/a605a675-9444-4017-8b34-d66be5b18c95}
}
\usage{
data(ArdecheStMartin)
}
\description{
A time series of simulated river discharge of the Ardeche river close
to its confluence with the Rhone (longitude = 4.658 \ latitude = 44.258) from
1951 to 2020. Time series extracted from the HERA dataset:
https://data.jrc.ec.europa.eu/dataset/a605a675-9444-4017-8b34-d66be5b18c95.
The Ardèche is Mediterranean river mostly known for tourism due
to its scenic gorges, but floods and droughts can impact the
local economy and environment.
}
\keyword{datasets}
