% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DanubeVienna}
\alias{DanubeVienna}
\title{Simulated river discharge of the Danube river at Vienna}
\format{
Two column dataframe:
#' \itemize{
\item time (POSIXct timestamp with 6-hourly resolution)
\item Q (6-houry mean discharge in cubic meter per second)
}
}
\source{
\url{https://data.jrc.ec.europa.eu/dataset/a605a675-9444-4017-8b34-d66be5b18c95}
}
\usage{
data(DanubeVienna)
}
\description{
A time series of simulated river discharge of the Danube river in Vienna
(longitude = 16.64 \ latitude = 48.13) from 1951 to 2020.
Time series extracted from the HERA dataset:
https://data.jrc.ec.europa.eu/dataset/a605a675-9444-4017-8b34-d66be5b18c95.
The Danube is the longest river in the EU and is an important waterway for
international trade, connecting several countries in Central and Eastern Europe.
}
\keyword{datasets}
