% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_functions.R
\name{findMMax}
\alias{findMMax}
\title{Find the index of the monthly maximum value in a given subset of a vector.}
\usage{
findMMax(indxs)
}
\arguments{
\item{indxs}{A numeric vector specifying the subset of indices.}
}
\value{
The index of the maximum value within the specified subset.
}
\description{
This function takes a vector and a subset of indices and returns the index of the maximum value within that subset.
}
\examples{
srs <- c(10, 20, 30, 40, 50)
findMMax(c(1, 3, 5))
# Output: 5
}
