% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_functions.R
\name{initPercentiles}
\alias{initPercentiles}
\title{Initialize Percentiles}
\usage{
initPercentiles(subsrs, percentM, percent, percentP)
}
\arguments{
\item{subsrs}{The input dataset.}

\item{percentM}{The percentile for the lower bound.}

\item{percent}{The percentile for the middle bound.}

\item{percentP}{The percentile for the upper bound.}
}
\value{
A list containing the calculated percentiles and probabilities.
}
\description{
This function calculates the percentiles and probabilities for a given dataset.
}
\examples{
subsrs <- c(1, 2, 3, 4, 5)
initPercentiles(subsrs, 25, 50, 75)
}
\seealso{
\code{\link[=tsEvaNanRunningPercentiles]{tsEvaNanRunningPercentiles()}}
}
