% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{max_daily_value}
\alias{max_daily_value}
\title{Max Daily Value Function}
\usage{
max_daily_value(timeseries)
}
\arguments{
\item{timeseries}{A time series with a 6-hourly resolution.}
}
\value{
A data frame containing the daily maximum values.
}
\description{
This function converts a 6-hourly time series to a daily time series and
calculates the maximum value for each day.
}
\examples{
# Example usage:
timeseries <- ArdecheStMartin
max_daily_value(timeseries)
}
