% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RhoneLyon}
\alias{RhoneLyon}
\title{Simulated river discharge of the Rhone river at Lyon}
\format{
Two column dataframe:
#' \itemize{
\item time (POSIXct timestamp with 6-hourly resolution)
\item Q (6-houry mean discharge in cubic meter per second)
}
}
\source{
\url{https://data.jrc.ec.europa.eu/dataset/a605a675-9444-4017-8b34-d66be5b18c95}
}
\usage{
data(RhoneLyon)
}
\description{
A time series of simulated river discharge of the Rhone river close
to its confluence with the Saone (longitude = 4.891 \ latitude = 45.772) from
1951 to 2020. Time series extracted from the HERA dataset:
https://data.jrc.ec.europa.eu/dataset/a605a675-9444-4017-8b34-d66be5b18c95.
The Rhône is France's most powerful river,characterized by a significant
seasonal variation in flow rates. The Rhône River is crucial for
transportation, hydropower generation, and irrigation in the region.
}
\keyword{datasets}
