% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_functions.R
\name{tsEstimateAverageSeasonality}
\alias{tsEstimateAverageSeasonality}
\title{Estimate Average Seasonality}
\usage{
tsEstimateAverageSeasonality(
  timeStamps,
  seasonalitySeries,
  timeWindow,
  peaks = T
)
}
\arguments{
\item{timeStamps}{The time stamps of the time series.}

\item{seasonalitySeries}{The series representing the seasonality.}

\item{timeWindow}{The time window used for averaging the seasonality.}

\item{peaks}{A logical value indicating whether to consider peaks in the seasonality.}
}
\value{
A list containing the estimated regime and the seasonality series.
\itemize{
\item \code{regime}: The estimated regime of the time series.
\item \code{Seasonality}: A data frame containing the average and varying seasonality series.
\itemize{
\item \code{averageSeasonalitySeries}: The average seasonality series.
\item \code{varyingSeasonalitySeries}: The varying seasonality series.
}
}
}
\description{
This function estimates the average seasonality of a time series based on the given parameters.
}
\examples{
timeStamps <- seq(as.Date("2022-01-01"), as.Date("2022-12-31"), by = "day")
seasonalitySeries <- sin(2 * pi * seq(0, 1, length.out = length(timeStamps)))
result <- tsEstimateAverageSeasonality(timeStamps, seasonalitySeries, timeWindow = 30, peaks = TRUE)
plot(result$regime, type = "l", xlab = "Day", ylab = "Regime", main = "Estimated Regime")
plot(result$Seasonality$averageSeasonalitySeries, type = "l", xlab = "Day", ylab = "Seasonality", main = "Average Seasonality")
plot(result$Seasonality$varyingSeasonalitySeries, type = "l", xlab = "Day", ylab = "Seasonality", main = "Varying Seasonality")

}
