% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_functions.R
\docType{methods}
\name{tsEvaComputeReturnLevelsGEV}
\alias{tsEvaComputeReturnLevelsGEV}
\title{tsEvaComputeReturnLevelsGEV}
\usage{
tsEvaComputeReturnLevelsGEV(
  epsilon,
  sigma,
  mu,
  epsilonStdErr,
  sigmaStdErr,
  muStdErr,
  returnPeriodsInDts
)
}
\arguments{
\item{epsilon}{The shape parameter of the GEV distribution.}

\item{sigma}{The scale parameter of the GEV distribution.}

\item{mu}{The location parameter of the GEV distribution.}

\item{epsilonStdErr}{The standard error of the shape parameter.}

\item{sigmaStdErr}{The standard error of the scale parameter.}

\item{muStdErr}{The standard error of the location parameter.}

\item{returnPeriodsInDts}{The return periods expressed in the corresponding time unit (e.g., years for yearly maxima).}
}
\value{
A list containing the following components:
\itemize{
\item \code{returnLevels}: A matrix of return levels corresponding to the specified return periods.
\item \code{returnLevelsErr}: A matrix of standard errors for the return levels.
}
}
\description{
This function calculates the return levels for a Generalized Extreme Value (GEV)
distribution given the GEV parameters and their standard errors.
The return periods are specified in a time unit that corresponds
to the size of the time segments for evaluating the maxima.
}
\references{
Stuart Coles (2001). \emph{An Introduction to Statistical Modeling of Extreme Values}. Springer.
Mentaschi, L., Vousdoukas, M., Voukouvalas, E., Sartini, L., Feyen, L.,
Besio, G., and Alfieri, L. (2016). The transformed-stationary approach:
a generic and simplified methodology for non-stationary extreme value analysis.
\emph{Hydrology and Earth System Sciences}, \strong{20}(9), 3527-3547.
doi:10.5194/hess-20-3527-2016.
}
\seealso{
\code{\link{empdis}}
}
\keyword{GEV,}
\keyword{computation,}
\keyword{errors}
\keyword{function,}
\keyword{levels,}
\keyword{return}
\keyword{standard}
