% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_functions.R
\docType{methods}
\name{tsEvaComputeReturnLevelsGEVFromAnalysisObj}
\alias{tsEvaComputeReturnLevelsGEVFromAnalysisObj}
\title{tsEvaComputeReturnLevelsGEVFromAnalysisObj}
\usage{
tsEvaComputeReturnLevelsGEVFromAnalysisObj(
  nonStationaryEvaParams,
  returnPeriodsInYears,
  ...
)
}
\arguments{
\item{nonStationaryEvaParams}{The parameters obtained from a non-stationary extreme value analysis.}

\item{returnPeriodsInYears}{The return periods expressed in years.}

\item{...}{Additional arguments.}
}
\value{
A list containing the following components:
\itemize{
\item \code{returnLevels}: A matrix of return levels corresponding to the specified return periods.
\item \code{returnLevelsErr}: A matrix of standard errors for the return levels.
\item \code{returnLevelsErrFit}: A matrix of standard errors for the return levels obtained from fitting the non-stationary model.
\item \code{returnLevelsErrTransf}: A matrix of standard errors for the return levels obtained from the transformed data.
}
}
\description{
This function calculates the return levels for a Generalized Extreme Value (GEV) distribution using the parameters obtained from a non-stationary extreme value analysis. It supports non-stationary analysis by considering different parameters for each time index.
}
\seealso{
\code{\link{tsEvaComputeReturnLevelsGEV}}

Other MyPackage functions: 
\code{\link{tsEvaComputeReturnLevelsGPD}()},
\code{\link{tsEvaComputeReturnLevelsGPDFromAnalysisObj}()},
\code{\link{tsEvaComputeReturnPeriodsGEV}()}
}
\concept{MyPackage functions}
