% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_functions.R
\docType{methods}
\name{tsEvaComputeReturnLevelsGPDFromAnalysisObj}
\alias{tsEvaComputeReturnLevelsGPDFromAnalysisObj}
\title{tsEvaComputeReturnLevelsGPDFromAnalysisObj}
\usage{
tsEvaComputeReturnLevelsGPDFromAnalysisObj(
  nonStationaryEvaParams,
  returnPeriodsInYears,
  ...
)
}
\arguments{
\item{nonStationaryEvaParams}{The non-stationary parameters obtained from the analysis object.}

\item{returnPeriodsInYears}{The return periods for which to compute the return levels, expressed in years.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A list with the following components:
\itemize{
\item \code{returnLevels} A vector of return levels corresponding to the specified return periods.
\item \code{returnLevelsErrFit} A vector of standard errors for the return levels estimated based on the fit.
\item \code{returnLevelsErrTransf} A vector of standard errors for the return levels estimated based on the transformed parameters.
}
}
\description{
\code{tsEvaComputeReturnLevelsGPDFromAnalysisObj} is a function that calculates the return levels for a Generalized Pareto Distribution (GPD) using the parameters obtained from an analysis object. It takes into account non-stationarity by considering time-varying parameters and their associated standard errors.
}
\seealso{
\code{\link{tsEvaComputeReturnLevelsGPD}}, \code{\link{tsEasyParseNamedArgs}}

Other MyPackage functions: 
\code{\link{tsEvaComputeReturnLevelsGEVFromAnalysisObj}()},
\code{\link{tsEvaComputeReturnLevelsGPD}()},
\code{\link{tsEvaComputeReturnPeriodsGEV}()}
}
\concept{MyPackage functions}
