% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_functions.R
\name{tsEvaComputeTimeRP}
\alias{tsEvaComputeTimeRP}
\title{tsEvaComputeTimeRP Function Documentation}
\usage{
tsEvaComputeTimeRP(params, RPiGEV, RPiGPD)
}
\arguments{
\item{params}{A data frame containing the following parameters:
\itemize{
\item epsilonGEV: Shape parameter for the Generalized Extreme Value (GEV) distribution.
\item muGEV: Location parameter for the GEV distribution.
\item sigmaGEV: Scale parameter for the GEV distribution.
\item epsilonGPD: Shape parameter for the Generalized Pareto (GPD) distribution.
\item thresholdGPD: Threshold parameter for the GPD distribution.
\item sigmaGPD: Scale parameter for the GPD distribution.
\item nPeaks: Number of peaks in the sample time horizon.
\item SampleTimeHorizon: Total sample time horizon.
}}

\item{RPiGEV}{Value of RP for the GEV distribution.}

\item{RPiGPD}{Value of RP for the GPD distribution.}
}
\value{
A vector with the calculated return period for GEV and GPD distributions.
}
\description{
This function calculates the return period of a given event for GEV and GPD distributions
at a given time index.
}
\examples{
# Example usage:
params <- data.frame(epsilonGEV = 0.2, muGEV = 3, sigmaGEV = 1,
                   epsilonGPD = 0.5, thresholdGPD = 2, sigmaGPD = 0.5,
                   nPeaks = 5, SampleTimeHorizon = 10)
RPcalc(params, RPiGEV = 10, RPiGPD = 20)
}
