% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{tsEvaPlotAllRLevelsGPD}
\alias{tsEvaPlotAllRLevelsGPD}
\title{tsEvaPlotAllRLevelsGPD}
\usage{
tsEvaPlotAllRLevelsGPD(
  nonStationaryEvaParams,
  stationaryTransformData,
  rlvmax,
  timeIndex,
  timeStamps,
  tstamps,
  trans,
  varargin
)
}
\arguments{
\item{nonStationaryEvaParams}{A list of non-stationary evaluation parameters
containing the GPD distribution parameters (epsilon, sigma, threshold),
time horizon start and end (thStart, thEnd), time horizon in years
(timeHorizonInYears), and number of peaks (nPeaks).}

\item{stationaryTransformData}{The stationary transformed data used for
the analysis.}

\item{rlvmax}{The maximum return level data, including the return periods
(haz.RP) and the actual return levels (QNS).}

\item{timeIndex}{The index of the time step used for analysis.}

\item{timeStamps}{The timestamps corresponding to the time steps in
the analysis.}

\item{tstamps}{The timestamps used for labeling the plot.}

\item{trans}{The transformation used to fit the EVD, either
"ori" (original) or "rev" (reverse).}

\item{...}{Additional optional arguments for customizing the plot.}
}
\value{
A plot object showing the relationship between return periods and
return levels for the GPD distribution at different timesteps.
}
\description{
This function generates a plot of return levels for a Generalized Pareto
Distribution (GPD) based on the provided parameters and data.
The plot showcases the evolving relationship between return periods and
return levels, allowing for visual analysis of extreme events
and their probabilities.
}
\examples{
tsEvaPlotAllRLevelsGPD(nonStationaryEvaParams, stationaryTransformData,
rlvmax, timeIndex, timeStamps, tstamps, mode = "inv", ylim = c(0, 100),
 ax = NULL)

}
\seealso{
\code{\link{tsEvaComputeReturnLevelsGPD}}
}
