% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{tsEvaPlotGEVImageSc}
\alias{tsEvaPlotGEVImageSc}
\title{Generate a GEV (Generalized Extreme Value) plot with raster image}
\usage{
tsEvaPlotGEVImageSc(
  Y,
  timeStamps,
  serix,
  epsilon,
  sigma,
  mu,
  returnPeriodInDts,
  maxObs,
  mode,
  varargin
)
}
\arguments{
\item{Y}{A vector of extreme values.}

\item{timeStamps}{A vector of timestamps corresponding to the extreme values.}

\item{serix}{The y-value at which to draw a horizontal line on the plot.}

\item{epsilon}{A numeric value representing the shape parameter of the GEV distribution.}

\item{sigma}{A vector of scale parameters corresponding to the timestamps.}

\item{mu}{A vector of location parameters corresponding to the timestamps.}

\item{returnPeriodInDts}{The return period in decimal time steps.}

\item{maxObs}{A data frame containing the maximum observations.}

\item{mode}{A character string indicating the mode of the plot. Possible values are "inv" (inverse) and "normal".}

\item{...}{Additional arguments to customize the plot.}
}
\value{
A ggplot object representing the GEV plot with a raster image.
}
\description{
This function generates a GEV plot with a raster image using the Generalized Extreme Value (GEV) distribution.
}
\examples{
# Generate a GEV plot with raster image
tsEvaPlotGEVImageSc(
  Y = c(1, 2, 3), timeStamps = c("2021-01-01", "2021-01-02", "2021-01-03"),
  serix = 2, epsilon = 0.1, sigma = c(0.5, 0.6, 0.7),
  mu = c(1, 2, 3), returnPeriodInDts = 1, maxObs = data.frame(time = "2021-01-01", value = 2),
  mode = "normal"
)

}
