% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{tsEvaPlotGEVImageScFromAnalysisObj}
\alias{tsEvaPlotGEVImageScFromAnalysisObj}
\title{tsEvaPlotGEVImageScFromAnalysisObj}
\usage{
tsEvaPlotGEVImageScFromAnalysisObj(
  Y,
  nonStationaryEvaParams,
  stationaryTransformData,
  trans,
  ...
)
}
\arguments{
\item{Y}{The input data.}

\item{nonStationaryEvaParams}{A list of non-stationary evaluation parameters.}

\item{stationaryTransformData}{The stationary transform data.}

\item{trans}{The transformation method.}

\item{...}{Additional arguments.}
}
\value{
The GEV image scatter plot.
}
\description{
\code{tsEvaPlotGEVImageScFromAnalysisObj}is a function that generates a GEV
(Generalized Extreme Value) time-varying distribution through time as
and show the evolution of exceedance probabilities.
}
\seealso{
\code{\link{tsEvaPlotGEVImageSc}}
}
