% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{tsEvaPlotGPDImageSc}
\alias{tsEvaPlotGPDImageSc}
\title{tsEvaPlotGPDImageSc}
\usage{
tsEvaPlotGPDImageSc(
  Y,
  timeStamps,
  serix,
  epsilon,
  sigma,
  threshold,
  peakplot,
  trans,
  varargin
)
}
\arguments{
\item{Y}{A vector of values.}

\item{timeStamps}{A vector of timestamps corresponding to the values.}

\item{serix}{A vector of series values.}

\item{epsilon}{A numeric value representing the shape parameter of the GPD.}

\item{sigma}{A vector of standard deviation values.}

\item{threshold}{A vector of threshold values.}

\item{peakplot}{A data frame containing peak values and their corresponding timestamps.}

\item{trans}{A character string indicating the transformation to be applied to the data.}

\item{varargin}{Additional optional arguments.}
}
\value{
A ggplot object representing the GPD plot.
}
\description{
\code{tsEvaPlotGPDImageSc}is a function that generates a  time series plot
of the Generalized Pareto Distribution (GPD) with evolving parameters
using the provided data.
}
\seealso{
\code{\link{tsEvaPlotGPDImageScFromAnalysisObj}}
}
