% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{tsEvaPlotGPDImageScFromAnalysisObj}
\alias{tsEvaPlotGPDImageScFromAnalysisObj}
\title{Plot GPD Image Score from Analysis Object}
\usage{
tsEvaPlotGPDImageScFromAnalysisObj(
  Y,
  nonStationaryEvaParams,
  stationaryTransformData,
  trans,
  ...
)
}
\arguments{
\item{Y}{The input data.}

\item{nonStationaryEvaParams}{A list containing non-stationary evaluation parameters.}

\item{stationaryTransformData}{A data frame containing stationary transform data.}

\item{trans}{The transformation method to be applied to the data.}

\item{...}{Additional arguments to be passed to the \code{\link{tsEvaPlotGPDImageSc}} function.}
}
\value{
The plot object.
}
\description{
This function plots the GPD (Generalized Pareto Distribution) image score from an analysis object.
}
\details{
This function takes the input data \code{Y}, non-stationary evaluation parameters \code{nonStationaryEvaParams},
stationary transform data \code{stationaryTransformData}, transformation method \code{trans}, and additional arguments \code{...}.
It then updates the arguments with the passed-in values, calculates the time stamps, and performs necessary transformations.
Finally, it plots the GPD image score using the \code{\link{tsEvaPlotGPDImageSc}} function and returns the plot object.
}
\examples{
# Example usage of tsEvaPlotGPDImageScFromAnalysisObj
tsEvaPlotGPDImageScFromAnalysisObj(Y, nonStationaryEvaParams, stationaryTransformData, trans)

}
\seealso{
\code{\link{tsEvaPlotGPDImageSc}}

\code{\link{stationaryTransformData}}

\code{\link{nonStationaryEvaParams}}
}
