% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{tsEvaPlotReturnLevelsGEV}
\alias{tsEvaPlotReturnLevelsGEV}
\title{tsEvaPlotReturnLevelsGEV}
\usage{
tsEvaPlotReturnLevelsGEV(
  epsilon,
  sigma,
  mu,
  epsilonStdErr,
  sigmaStdErr,
  muStdErr,
  rlvmax,
  tstamps,
  trans,
  ...
)
}
\arguments{
\item{epsilon}{The shape parameter of the GEV distribution.}

\item{sigma}{The scale parameter of the GEV distribution.}

\item{mu}{The location parameter of the GEV distribution.}

\item{epsilonStdErr}{The standard error of the shape parameter.}

\item{sigmaStdErr}{The standard error of the scale parameter.}

\item{muStdErr}{The standard error of the location parameter.}

\item{rlvmax}{A data frame containing the return levels of annual maxima.}

\item{tstamps}{The title for the plot.}

\item{trans}{The transformation used to fit the EVD, either "ori" (original)
or "rev" (reverse). "inv" and "lninv" are also available
but in development phase.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A ggplot object representing the plot of return levels.
}
\description{
\code{tsEvaPlotReturnLevelsGEV} is a function that plots the return levels
using the Generalized Extreme Value (GEV) distribution.
}
\seealso{
\code{\link{tsEvaComputeReturnLevelsGEV}}
\code{\link{tsEvaPlotReturnLevelsGEVFromAnalysisObj}}
}
