% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{tsEvaPlotReturnLevelsGPD}
\alias{tsEvaPlotReturnLevelsGPD}
\title{tsEvaPlotReturnLevelsGPD}
\usage{
tsEvaPlotReturnLevelsGPD(
  epsilon,
  sigma,
  threshold,
  epsilonStdErr,
  sigmaStdErr,
  thresholdStdErr,
  nPeaks,
  timeHorizonInYears,
  rlvmax,
  tstamps,
  trans,
  ...
)
}
\arguments{
\item{epsilon}{The shape parameter of the GPD.}

\item{sigma}{The scale parameter of the GPD.}

\item{threshold}{The threshold parameter of the GPD.}

\item{epsilonStdErr}{The standard error of the shape parameter.}

\item{sigmaStdErr}{The standard error of the scale parameter.}

\item{thresholdStdErr}{The standard error of the threshold parameter.}

\item{nPeaks}{The number of peaks used in the GPD estimation.}

\item{timeHorizonInYears}{The time horizon in years for the GPD estimation.}

\item{rlvmax}{A data frame containing the return levels of annual maxima.}

\item{tstamps}{The title for the plot.}

\item{trans}{The transformation type for the return levels.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A ggplot object representing the plot of return levels.
}
\description{
\code{tsEvaPlotReturnLevelsGPD} is a function that plots the return levels
using the Generalized Pareto Distribution (GPD).
}
\seealso{
\code{\link{tsEvaComputeReturnLevelsGPD}}
\code{\link{tsEvaPlotReturnLevelsGPDFromAnalysisObj}}
}
