% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{tsEvaRLCurveSummary}
\alias{tsEvaRLCurveSummary}
\title{tsEvaRLCurveSummary Function}
\usage{
tsEvaRLCurveSummary(nonStationaryEvaParams, stationaryTransformData, trans)
}
\arguments{
\item{nonStationaryEvaParams}{A list containing the parameters for non-stationary evaluation.}

\item{stationaryTransformData}{The transformed stationary data.}

\item{trans}{The transformation used to fit the EVD, either "ori" (original) or "rev" (reverse).}
}
\value{
A data frame containing the return periods and the corresponding highest, lowest, and median return levels.
}
\description{
This function provides a summary of the return level curves for a non-stationary time series.
It retains the highest, lowest and median RL curves.
}
\examples{
tsEvaRLCurveSummary(nonStationaryEvaParams, stationaryTransformData, trans)
}
