% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_functions.R
\name{tsEvaTransformSeriesToStationaryMultiplicativeSeasonality}
\alias{tsEvaTransformSeriesToStationaryMultiplicativeSeasonality}
\title{Transform Series to Stationary with Multiplicative Seasonality}
\usage{
tsEvaTransformSeriesToStationaryMultiplicativeSeasonality(
  timeStamps,
  series,
  timeWindow,
  seasonalityVar = T,
  ...
)
}
\arguments{
\item{timeStamps}{A vector of timestamps for the time series data.}

\item{series}{A vector of the time series data.}

\item{timeWindow}{The size of the moving window used for trend estimation.}

\item{seasonalityVar}{A logical value indicating whether to consider seasonality in the transformation. Default is TRUE.}

\item{...}{Additional arguments to be passed to other functions.}
}
\value{
A list containing the transformed data and various statistics and errors.
\describe{
\item{runningStatsMulteplicity}{The size of the moving window used for trend estimation.}
\item{stationarySeries}{The transformed stationary series.}
\item{trendSeries}{The trend component of the transformed series.}
\item{trendSeriesNonSeasonal}{The trend component of the original series without seasonality.}
\item{stdDevSeries}{The standard deviation component of the transformed series.}
\item{stdDevSeriesNonSeasonal}{The standard deviation component of the original series without seasonality.}
\item{trendNonSeasonalError}{The error on the non-seasonal trend component.}
\item{stdDevNonSeasonalError}{The error on the non-seasonal standard deviation component.}
\item{trendSeasonalError}{The error on the seasonal trend component.}
\item{stdDevSeasonalError}{The error on the seasonal standard deviation component.}
\item{trendError}{The overall error on the trend component.}
\item{stdDevError}{The overall error on the standard deviation component.}
\item{Regime}{The estimated regime of the trend seasonality.}
\item{timeStamps}{The input timestamps.}
\item{nonStatSeries}{The original non-stationary series.}
\item{statSer3Mom}{The third moment of the transformed stationary series.}
\item{statSer4Mom}{The fourth moment of the transformed stationary series.}
}
}
\description{
This function decomposes a time series into a season-dependent trend and a season-dependent standard deviation.
It performs a transformation from non-stationary to stationary.
}
\examples{
timeStamps <- seq(as.Date("2022-01-01"), as.Date("2022-12-31"), by = "day")
series <- sin(2 * pi * seq(0, 1, length.out = length(timeStamps)))
transformedData <- tsEvaTransformSeriesToStationaryMultiplicativeSeasonality(timeStamps, series, 30)
}
