% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_functions.R
\name{tsEvaTransformSeriesToStationaryPeakTrend}
\alias{tsEvaTransformSeriesToStationaryPeakTrend}
\title{Transform Time Series to Stationary only using extremes}
\usage{
tsEvaTransformSeriesToStationaryPeakTrend(
  timeStamps,
  series,
  timeWindow,
  TrendTh,
  ...
)
}
\arguments{
\item{timeStamps}{A vector of time stamps corresponding to the observations in the series.}

\item{series}{A vector of the time series data.}

\item{timeWindow}{The size of the time window used for detrending.}

\item{TrendTh}{The threshold for fitting the trend on the means above a certain quantile. Default is 0.5.}

\item{...}{Additional arguments to be passed to other functions.}
}
\value{
A list containing the following components:
\itemize{
\item \code{runningStatsMulteplicity}: The multiplicity of running statistics.
\item \code{stationarySeries}: The stationary series after removing the trend.
\item \code{trendSeries}: The trend component of the series.
\item \code{trendSeriesNonSeasonal}: NULL (not used).
\item \code{trendError}: The error on the trend component.
\item \code{stdDevSeries}: The standard deviation series.
\item \code{stdDevSeriesNonSeasonal}: NULL (not used).
\item \code{stdDevError}: The error on the standard deviation series.
\item \code{timeStamps}: The time stamps.
\item \code{nonStatSeries}: The original non-stationary series.
\item \code{statSer3Mom}: The running mean of the third moment of the stationary series.
\item \code{statSer4Mom}: The running mean of the fourth moment of the stationary series.
}
}
\description{
This function transforms a time series to a stationary one by focusing on extremes.
The trend and slowly varying amplitude are computed on values above a
threshold defined by the user.
}
\examples{
timeStamps <- seq(as.Date("2022-01-01"), as.Date("2022-01-31"), by = "day")
series <- rnorm(length(timeStamps))
transformedData <- tsEvaTransformSeriesToStationaryPeakTrend(timeStamps, series, timeWindow = 7)

}
\seealso{
\code{\link[=tsEvaFindTrendThreshold]{tsEvaFindTrendThreshold()}}
}
