% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_functions.R
\name{tsEvaTransformSeriesToStatSeasonal_ciPercentile}
\alias{tsEvaTransformSeriesToStatSeasonal_ciPercentile}
\title{tsEvaTransformSeriesToStatSeasonal_ciPercentile}
\usage{
tsEvaTransformSeriesToStatSeasonal_ciPercentile(
  timeStamps,
  series,
  timeWindow,
  percentile
)
}
\arguments{
\item{timeStamps}{A vector of time stamps for the time series.}

\item{series}{The original time series.}

\item{timeWindow}{The length of the moving window used for trend estimation.}

\item{percentile}{The percentile value used for computing the slowly varying percentile.}
}
\value{
A list containing the following components:
\itemize{
\item{runningStatsMulteplicity: The size of each sample used to compute the average.}
\item{stationarySeries: The transformed stationary series.}
\item{trendSeries: The trend series.}
\item{trendSeriesNonSeasonal: The non-seasonal trend series.}
\item{stdDevSeries: The season-dependent standard deviation series.}
\item{stdDevSeriesNonSeasonal: The non-seasonal standard deviation series.}
\item{trendError: The error on the trend.}
\item{stdDevError: The error on the standard deviation.}
\item{statSer3Mom: The 3rd moment of the transformed stationary series.}
\item{statSer4Mom: The 4th moment of the transformed stationary series.}
\item{nonStatSeries: The original non-stationary series.}
\item{Regime: The regime of the trend seasonality.}
\item{timeStamps: The time stamps.}
\item{trendNonSeasonalError: The error on the non-seasonal trend.}
\item{stdDevNonSeasonalError: The error on the non-seasonal standard deviation.}
\item{trendSeasonalError: The error on the seasonal trend.}
\item{stdDevSeasonalError: The error on the seasonal standard deviation.}
}
}
\description{
This function decomposes a time series into a season-dependent trend and a season-dependent standard deviation.
The season-dependent amplitude is given by a seasonal factor multiplied by a slowly varying percentile.
}
\section{this function decomposes the series into a season-dependent trend and a}{
season-dependent standard deviation.
The season-dependent standard deviation is given by a seasonal factor
multiplied by a slowly varying standard deviation.
transformation non stationary -> stationary
x(t) =  (y(t) - trend(t) - ssn_trend(t))/(stdDev(t)*ssn_stdDev(t))
transformation stationary -> non stationary
y(t) = stdDev(t)*ssn_stdDev(t)*x(t) + trend(t) + ssn_trend(t)
trasfData.trendSeries = trend(t) + ssn_trend(t)
trasfData.stdDevSeries = stdDev(t)*ssn_stdDev(t)
}

\examples{
timeAndSeries <- ArdecheStMartin
timeStamps <- ArdecheStMartin[,1]
series <- ArdecheStMartin[,2]
timeWindow <- 30*365 # 30 years
percentile <- 90
result <- tsEvaTransformSeriesToStatSeasonal_ciPercentile(timeStamps,
series, timeWindow, percentile, seasonalityVar=F)
plot(result$trendSeries)
}
