% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_functions.R
\name{computeMonthlyMaxima}
\alias{computeMonthlyMaxima}
\title{Compute Monthly Maxima}
\usage{
computeMonthlyMaxima(timeAndSeries)
}
\arguments{
\item{timeAndSeries}{A data frame containing the time stamps and series values.
The first column should contain the time stamps, and the second column should contain the series values.}
}
\value{
A list containing the monthly maxima, corresponding dates, and indices.
- \code{monthlyMax}: A vector of the monthly maximum values.
- \code{monthlyMaxDate}: A vector of the dates corresponding to the monthly maximum values.
- \code{monthlyMaxIndx}: A vector of the indices of the monthly maximum values in the original series.
}
\description{
This function computes the monthly maxima of a time series.
}
\examples{
timeAndSeries <- data.frame(timeStamps = c("2022-01-01", "2022-01-02", "2022-02-01", "2022-02-02"),
                            series = c(10, 15, 5, 20))
computeMonthlyMaxima(timeAndSeries)

}
