% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_functions.R
\name{computeMonthlyMaxima}
\alias{computeMonthlyMaxima}
\title{computeMonthlyMaxima}
\usage{
computeMonthlyMaxima(timeAndSeries)
}
\arguments{
\item{timeAndSeries}{A data frame containing the time stamps and series values.
The first column should contain the time stamps,
and the second column should contain the series values.}
}
\value{
A list containing the monthly maxima, corresponding dates, and indices.
\describe{
\item{\code{monthlyMax}}{A vector of the monthly maximum values.}
\item{\code{monthlyMaxDate}}{A vector of the dates corresponding to the monthly maximum values.}
\item{\code{monthlyMaxIndx}}{A vector of the indices of the monthly maximum values in the original series.}
}
}
\description{
\code{computeMonthlyMaxima} is a function that computes the monthly maxima
of a time series.
}
\examples{
timeAndSeries <- ArdecheStMartin
computeMonthlyMaxima(timeAndSeries)
}
