% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_functions.R
\name{empdis}
\alias{empdis}
\title{Empirical Distribution Function}
\usage{
empdis(x, nyr)
}
\arguments{
\item{x}{A numeric vector representing the dataset.}

\item{nyr}{An integer representing the number of years in the dataset.}
}
\value{
A data frame containing the empirical return periods, hazard return periods,
Cunnane return periods, Gumbel values, empirical probabilities, hazard probabilities,
Cunnane probabilities, the original dataset, and the timestamp.
}
\description{
This function calculates the empirical distribution function for a given dataset.
}
\examples{
x <- c(1, 2, 3, 4, 5)
nyr <- 5
empdis(x, nyr)

}
