% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_functions.R
\name{initPercentiles}
\alias{initPercentiles}
\title{Initialize Percentiles}
\usage{
initPercentiles(subsrs, percentM, percent, percentP)
}
\arguments{
\item{subsrs}{The input dataset.}

\item{percentM}{The percentile for the lower bound.}

\item{percent}{The percentile for the middle bound.}

\item{percentP}{The percentile for the upper bound.}
}
\value{
A list containing the calculated percentiles and probabilities.
}
\description{
This function calculates percentiles for a given dataset
}
\examples{
timeAndSeries <- ArdecheStMartin
timeStamps <- ArdecheStMartin[,1]
series <- ArdecheStMartin[,2]
initPercentiles(series, 89, 90, 91)
}
\seealso{
\code{\link[=tsEvaNanRunningPercentiles]{tsEvaNanRunningPercentiles()}}
}
