% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_functions.R
\docType{methods}
\name{tsEvaComputeReturnPeriodsGPD}
\alias{tsEvaComputeReturnPeriodsGPD}
\title{tsEvaComputeReturnPeriodsGPD}
\usage{
tsEvaComputeReturnPeriodsGPD(
  epsilon,
  sigma,
  threshold,
  peaks,
  nPeaks,
  peaksID,
  sampleTimeHorizon
)
}
\arguments{
\item{epsilon}{The shape parameter of the GPD.}

\item{sigma}{The scale parameter of the GPD.}

\item{threshold}{The threshold value for the GPD.}

\item{peaks}{A vector containing the peak values.}

\item{nPeaks}{The number of peak values.}

\item{peaksID}{An identifier for each peak value.}

\item{sampleTimeHorizon}{The time horizon of the sample.}
}
\value{
A list containing the following components:
\itemize{
\item \code{GpdPseudo}: A matrix of pseudo observations obtained from the GPD for each peak value at every time step.
\item \code{returnPeriods}: A matrix of return periods corresponding to the pseudo observations.
\item \code{PseudoObs}: A data frame containing the pseudo observations and their corresponding identifiers.
}
}
\description{
This function computes the return periods of a set of observations (peaks) for a
Generalized Pareto Distribution (GPD), given the GPD parameters,
threshold, peaks data, and sample time horizon.
}
\references{
Mentaschi, L., Vousdoukas, M., Voukouvalas, E., Sartini, L., Feyen, L., Besio, G., and Alfieri, L. (2016). The transformed-stationary approach: a generic and simplified methodology for non-stationary extreme value analysis. \emph{Hydrology and Earth System Sciences}, \strong{20}(9), 3527-3547. doi:10.5194/hess-20-3527-2016.
}
\seealso{
\code{\link{empdis}}
}
\keyword{GPD,}
\keyword{computation,}
\keyword{function,}
\keyword{observations}
\keyword{periods,}
\keyword{pseudo}
\keyword{return}
