% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainTSEVA.R
\name{TsEvaNs}
\alias{TsEvaNs}
\title{TsEvaNs Function}
\usage{
TsEvaNs(timeAndSeries, timeWindow, ...)
}
\arguments{
\item{timeAndSeries}{A data frame containing the timestamp and corresponding series data.}

\item{timeWindow}{The time window for analysis.}

\item{...}{Additional arguments for customization.}

\item{transfType}{The transformation type for non-stationary EVA. It can be one of the following:
\itemize{trend: Long-term variations of the timeseries.
seasonal: Long-term and seasonal variations of extremes.
trendCIPercentile: Long-term variations of extremes using a specified percentile.
trendPeak: Long-term variations of the peaks.}}

\item{minPeakDistanceInDays}{The minimum peak distance in days.}

\item{seasonalityVar}{A logical value indicating whether to consider seasonality in the analysis.}

\item{potEventsPerYear}{The number of potential events per year.}

\item{minEventsPerYear}{The minimum number of events per year.}

\item{gevMaxima}{The type of maxima for the GEV distribution (annual or monthly).}

\item{ciPercentile}{The percentile value for confidence intervals.}

\item{gevType}{The type of GEV distribution (GEV or GPD).}

\item{evdType}{The type of extreme value distribution (GEV or GPD).}

\item{tail}{The mode of the analysis (e.g., high for flood peaks or low for drought peaks).}

\item{epy}{The average number of events per year, can be specified by the
user or automatically set according to the tail selected.}

\item{lowdt}{The temporal resolultion used for low values. default is 7 days.}
}
\value{
A list containing the results of the non-stationary EVA. Containing the following components:
\itemize{nonStationaryEvaParams: The estimated parameters for non-stationary EVA.
Parameters include GEV and GPD parameters for each timestep, confidence intervals, and other statistical measures.
stationaryTransformData: The transformed data for stationary EVA.Includes the stationary series, trend, and standard deviation series.}
}
\description{
This function performs non-stationary extreme value analysis (EVA) on a time series data.
}
\details{
The function takes a time series data and performs non-stationary EVA using various transformation types and parameters depending on the input data provided.
Results are returned as a list containing the non-stationary EVA parameters and the transformed data for stationary EVA and can be used as input for further analysis.
In particular for the following function
}
\examples{
# Example usage of TsEvaNs function
TimeAndSeries <- data.frame(timestamp = seq(as.POSIXct("2000-01-01"), as.POSIXct("2020-12-31"), by = "day",data = rnorm(7671))
timeWindow <- 365
result <- TsEvaNs(timeAndSeries, timeWindow, transfType = 'trendPeaks', ciPercentile = 90, mode = 'drought')

}
