% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{tsEvaPlotGEVImageScFromAnalysisObj}
\alias{tsEvaPlotGEVImageScFromAnalysisObj}
\title{Generate GEV Image Scatter Plot from Analysis Object}
\usage{
tsEvaPlotGEVImageScFromAnalysisObj(
  Y,
  nonStationaryEvaParams,
  stationaryTransformData,
  trans,
  ...
)
}
\arguments{
\item{Y}{The input data.}

\item{nonStationaryEvaParams}{A list of non-stationary evaluation parameters.}

\item{stationaryTransformData}{The stationary transform data.}

\item{trans}{The transformation method.}

\item{...}{Additional arguments.}
}
\value{
The GEV image scatter plot.
}
\description{
This function generates a GEV (Generalized Extreme Value) image scatter plot from an analysis object.
}
\examples{
# Example usage of tsEvaPlotGEVImageScFromAnalysisObj function
tsEvaPlotGEVImageScFromAnalysisObj(Y, nonStationaryEvaParams, stationaryTransformData, trans)

}
