% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\docType{methods}
\name{tsEvaPlotReturnLevelsGEVFromAnalysisObj}
\alias{tsEvaPlotReturnLevelsGEVFromAnalysisObj}
\title{tsEvaPlotReturnLevelsGEVFromAnalysisObj}
\usage{
tsEvaPlotReturnLevelsGEVFromAnalysisObj(
  nonStationaryEvaParams,
  stationaryTransformData,
  timeIndex,
  trans,
  ...
)
}
\arguments{
\item{nonStationaryEvaParams}{The non-stationary parameters obtained from the
analysis object.}

\item{stationaryTransformData}{The stationary transformed data obtained from
the analysis object.}

\item{timeIndex}{The index at which the time-varying analysis should be
estimated.}

\item{trans}{The transformation used to fit the EVD. Can be "ori" for no
transformation or "rev" for reverse transformation.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
Plot 1: RLtstep: return level curve with confidence interval for the
selected timeIndex
Plot 2: beam: beam of return level curve for all with highlited curve
for selected timeIndex
}
\description{
\code{tsEvaPlotReturnLevelsGEVFromAnalysisObj} is a function that plots the
return levels for a Generalized Extreme Value (GEV) distribution using the
parameters obtained from an analysis object. It considers non-stationarity
by considering time-varying parameters and their associated standard errors.
}
\references{
Mentaschi, L., Vousdoukas, M., Voukouvalas, E., Sartini, L., Feyen, L., Besio,
G., and Alfieri, L. (2016). The transformed-stationary approach: a generic
and simplified methodology for non-stationary extreme value analysis.
\emph{Hydrology and Earth System Sciences}, 20, 3527-3547.
doi:10.5194/hess-20-3527-2016.
}
