% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{tsEvaPlotReturnLevelsGPD}
\alias{tsEvaPlotReturnLevelsGPD}
\title{Plot Return Levels using Generalized Pareto Distribution (GPD)}
\usage{
tsEvaPlotReturnLevelsGPD(
  epsilon,
  sigma,
  threshold,
  epsilonStdErr,
  sigmaStdErr,
  thresholdStdErr,
  nPeaks,
  timeHorizonInYears,
  rlvmax,
  tstamps,
  trans,
  ...
)
}
\arguments{
\item{epsilon}{The shape parameter of the GPD.}

\item{sigma}{The scale parameter of the GPD.}

\item{threshold}{The threshold parameter of the GPD.}

\item{epsilonStdErr}{The standard error of the shape parameter.}

\item{sigmaStdErr}{The standard error of the scale parameter.}

\item{thresholdStdErr}{The standard error of the threshold parameter.}

\item{nPeaks}{The number of peaks used in the GPD estimation.}

\item{timeHorizonInYears}{The time horizon in years for the GPD estimation.}

\item{rlvmax}{A data frame containing the return levels of annual maxima.}

\item{tstamps}{The title for the plot.}

\item{trans}{The transformation type for the return levels.}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A ggplot object representing the plot of return levels.
}
\description{
This function plots the return levels using the Generalized Pareto Distribution (GPD).
}
\examples{
tsEvaPlotReturnLevelsGPD(
  epsilon = 0.1, sigma = 1, threshold = 0, epsilonStdErr = 0.05, sigmaStdErr = 0.1,
  thresholdStdErr = 0.05, nPeaks = 10, timeHorizonInYears = 50,
  rlvmax = data.frame(QNS = c(1, 2, 3), Qreal = c(10, 20, 30)),
  tstamps = "Return Levels", trans = "inv")
  @export
}
\seealso{
\code{\link{tsEvaComputeReturnLevelsGPD}} \code{\link{tsEvaPlorReturnLevelsGPDFromAnalysisObj}}
}
