% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{tsEvaPlotSeriesTrendStdDevFromAnalyisObj}
\alias{tsEvaPlotSeriesTrendStdDevFromAnalyisObj}
\title{Plot Time Series with Trend and Standard Deviation}
\usage{
tsEvaPlotSeriesTrendStdDevFromAnalyisObj(
  nonStationaryEvaParams,
  stationaryTransformData,
  mode,
  ...
)
}
\arguments{
\item{nonStationaryEvaParams}{The non-stationary evaluation parameters.}

\item{stationaryTransformData}{The stationary transformed data.}

\item{mode}{The mode of the plot (optional).}

\item{...}{Additional arguments to customize the plot (optional).}
}
\value{
A ggplot object representing the plot.
}
\description{
This function plots a time series along with its trend and standard deviation.
}
\examples{
# Example usage
tsEvaPlotSeriesTrendStdDevFromAnalyisObj(nonStationaryEvaParams, stationaryTransformData, mode = "inv")
}
\seealso{
\code{\link{trans}}
}
