% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{tsEvaPlotSeriesTrendStdDevFromAnalyisObj}
\alias{tsEvaPlotSeriesTrendStdDevFromAnalyisObj}
\title{tsEvaPlotSeriesTrendStdDevFromAnalyisObj}
\usage{
tsEvaPlotSeriesTrendStdDevFromAnalyisObj(
  nonStationaryEvaParams,
  stationaryTransformData,
  mode,
  ...
)
}
\arguments{
\item{nonStationaryEvaParams}{The non-stationary evaluation parameters.}

\item{stationaryTransformData}{The stationary transformed data.}

\item{mode}{The mode of the plot (optional).}

\item{...}{Additional arguments to customize the plot (optional).}
}
\value{
A ggplot object representing the plot.
}
\description{
\code{tsEvaPlotTrendStdDevFromAnalysisObj}is a function that plots a
time series along with its trend and standard deviation.
}
