% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eva_functions.R
\name{tsGetPOT}
\alias{tsGetPOT}
\title{tsGetPOT Function}
\usage{
tsGetPOT(
  ms,
  pcts,
  desiredEventsPerYear,
  minEventsPerYear,
  minPeakDistanceInDays,
  tail
)
}
\arguments{
\item{ms}{A matrix containing the time series data with two columns: the first column represents the time and the second column represents the values.}

\item{pcts}{A numeric vector specifying the percentiles to be used as thresholds for identifying peaks.}

\item{desiredEventsPerYear}{The desired number of events per year.}

\item{minEventsPerYear}{The minimum number of events per year.}

\item{minPeakDistanceInDays}{The minimum distance between two peaks in days.}

\item{tail}{The tail to be studied for POT selection, either 'high' or 'low'.}
}
\value{
A list containing the following fields:
\describe{
\item{threshold}{The threshold value used for identifying peaks.}
\item{thresholdError}{The error associated with the threshold value.}
\item{percentile}{The percentile value used as the threshold.}
\item{peaks}{The values of the identified peaks.}
\item{stpeaks}{The start indices of the identified peaks.}
\item{endpeaks}{The end indices of the identified peaks.}
\item{ipeaks}{The indices of the identified peaks.}
\item{time}{The time values corresponding to the identified peaks.}
\item{pars}{The parameters of the Generalized Pareto Distribution (GPD) fitted to the peaks.}
}
}
\description{
This function calculates the Peaks Over Threshold (POT) for a given time series data.
}
\examples{
# Create a sample time series data
time <- seq(as.Date("2000-01-01"), as.Date("2000-12-31"), by = "day")
values <- rnorm(length(time))
ms <- cbind(time, values)

# Calculate the POT using the tsGetPOT function
pcts <- c(90, 95, 99)
desiredEventsPerYear <- 5
minEventsPerYear <- 2
minPeakDistanceInDays <- 10
mode <- "flood"
POTdata <- tsGetPOT(ms, pcts, desiredEventsPerYear, minEventsPerYear, minPeakDistanceInDays, mode)

# Print the results
print(POTdata)

}
\seealso{
\code{\link[=tsSampleData]{tsSampleData()}}
}
